+++ src/typescript/src/domain/errors/DomainError.ts
